Выносит утилиты в модули и использует их в main.js
